Retirando warnings ao compilar para win64

+++ src/lib3270/host.c
@@ -570,8 +570,8 @@ static int do_connect(H3270 *hSession, const char *n)
 	/* Setup socket I/O. */
 //	add_input_calls(hSession,net_input,net_exception);
 #ifdef _WIN32
-	hSession->ns_exception_id	= AddExcept((int) hSession->sockEvent, hSession, net_exception);
-	hSession->ns_read_id		= AddInput((int) hSession->sockEvent, hSession, net_input);
+	hSession->ns_exception_id	= AddExcept(hSession->sockEvent, hSession, net_exception);
+	hSession->ns_read_id		= AddInput(hSession->sockEvent, hSession, net_input);
 #else
 	hSession->ns_exception_id	= AddExcept(hSession->sock, hSession, net_exception);
 	hSession->ns_read_id		= AddInput(hSession->sock, hSession, net_input);
